+++ Basic-Car-Maintenance/Shared/Models/AppIcon.swift
@@ -1,43 +1,4 @@
 import UIKit
-import Observation
-
-/// The ViewModel responsible for allowing users to change the AppIcon
-@Observable class ChangeAppIconViewModel {
-  private(set) var selectedAppIcon: AppIcon
-  
-  init() {
-    if let iconName = UIApplication.shared.alternateIconName,
-       let appIcon = AppIcon(rawValue: iconName) {
-      self._selectedAppIcon = appIcon
-    } else {
-      self._selectedAppIcon = .primary
-    }
-  }
-  
-  func updateAppIcon(to icon: AppIcon) {
-    let previousAppIcon = selectedAppIcon
-    selectedAppIcon = icon
-    
-    Task { @MainActor in
-      guard UIApplication.shared.alternateIconName != icon.iconName else {
-         print("No need to update icon since we're already using this icon.")
-        return
-      }
-      
-      do {
-        try await UIApplication.shared.setAlternateIconName(icon.iconName)
-        print("Successfully updated icon.")
-      } catch {
-        // We're only logging the error here and not actively handling the app icon failure
-        // since it's very unlikely to fail.
-        print("Updating icon to \(String(describing: icon.iconName)) failed.")
-        
-        // Restore previous app icon
-        selectedAppIcon = previousAppIcon
-      }
-    }
-  }
-}
 
 enum AppIcon: String, CaseIterable, Identifiable {
   case primary = "AppIcon"
@@ -48,6 +9,8 @@ enum AppIcon: String, CaseIterable, Identifiable {
   case carOrange = "AppIcon-car-orange"
   
   var id: String { rawValue }
+  
+  /// the name of the icon in the App Bundle.
   var iconName: String? {
     switch self {
     /// returns `nil`. Use this case to reset the app icon back to its primary icon.
@@ -58,6 +21,7 @@ enum AppIcon: String, CaseIterable, Identifiable {
     }
   }
   
+  /// A UI presentable string.
   var description: String {
     switch self {
     case .primary:
